requirements.txt
airdrop-ng==1.1
airgraph-ng==1.1
alabaster==0.7.13
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==2.15.6
asttokens==2.4.1
async-lru==2.0.4
attrs==22.2.0
autocommand==2.2.2
Automat==22.10.0
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.10.1
bleach==6.1.0
blinker==1.6.2
borgbackup==1.2.6
borgmatic==1.8.4
btrfsutil==6.6.1
CacheControl==0.13.1
cairocffi==1.6.1
capstone==4.0.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
constantly==15.1.0
contourpy==1.2.0
cryptography==41.0.5
cvxopt==1.3.2
cycler==0.12.1
cypari2==2.1.4
cysignals==1.11.4
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
docopt==0.6.2
docutils==0.20.1
executing==2.0.0
fastjsonschema==2.18.1
feedparser==6.0.10
filelock==3.12.4
Flask==2.3.3
fonttools==4.44.0
fpylll==0.6.0
fqdn==1.5.1
git-filter-repo==2.38.0
gmpy2==2.1.5
gpg==1.23.1
gps==3.25
greenlet==3.0.1
html2text==2020.1.16
html5lib==1.1
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
impacket==0.9.24
importlib-metadata==5.0.0
incremental==22.10.0
inflect==7.0.0
informant==0.5.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keyring==24.2.0
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
ldap3==2.9
lensfun==0.3.4
libfdt==1.7.0
libvirt-python==9.9.0
lit==16.0.6.dev0
lockfile==0.12.2
lrcalc==2.0.0
Mako==1.2.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
memory-allocator==0.1.3
meson==1.2.3
mistune==2.0.5
more-itertools==10.1.0
mpi4py==3.1.5
mpmath==1.3.0
msgpack==1.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.7
netsnmp-python==1.0a1
networkx==3.2.1
nftables==0.1
notebook_shim==0.2.3
numpy==1.26.1
open-fprintd==0.6
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pcapy-ng==1.0.5
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
ply==3.11
pooch==1.8.0
powerline-status==2.8.3
pplpy==0.8.9
primecountpy==0.1.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.5
pyasn1==0.4.8
pybind11==2.11.1
pycairo==1.25.1
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.12.0
pycurl==7.45.2
pydantic==2.4.2
pydantic_core==2.10.1
pygame==2.4.0
Pygments==2.16.1
PyGObject==3.46.0
pyinotify==0.9.6
pylint==2.17.7
pynvim==0.4.3
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypng==0.20231004.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
pyte==0.8.1
pytest==7.4.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-pam==2.0.2
python-snappy==0.6.1
python-validity==0.13
pytz==2023.3.post1
pyusb==1.2.1
pywayland==0.4.17
pywlroots==0.16.6
pyxdg==0.28
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
qtile==0.23.0
ranger-fm==1.9.3
referencing==0.30.2
Reflector==2023.6.28.0.36.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.6.0
rpds-py==0.10.6
rpy2==3.5.14
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.7
sagemath-bliss==10.1
sagemath-coxeter3==10.1
sagemath-mcqd==10.1
sagemath-meataxe==10.1
sagemath-sirocco==10.1
sagemath-standard==10.1
sagemath-tdlib==10.1
scipy==1.11.3
SecretStorage==3.3.3
Send2Trash==1.8.2
setproctitle==1.3.3
sgmllib3k==1.0.0
sh==1.14.3
six==1.16.0
smbus==1.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
speedtest-cli==2.1.3
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
sympy==1.12
syncthing-gtk==0.9.4.5
tabulate==0.9.0
TBB==0.2
testpath==0.6.0
thefuck==3.32
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.2
torbrowser-launcher==0.3.6
tornado==6.3
tqdm==4.66.1
traitlets==5.13.0
trash-cli==0.23.2.13.2
trove-classifiers==2023.10.27
Twisted==22.10.0
typing_extensions==4.8.0
tzlocal==5.2
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==1.26.15
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
volatility3==2.4.1
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==10.4
Werkzeug==2.3.7
wrapt==1.14.1
wxPython==4.2.1
xcffib==1.4.0
xkbcommon==0.8
XlsxWriter==3.1.0
yara-python==4.3.1
yt-dlp==2023.10.13
zipp==3.17.0
zope.interface==6.0
